add-canonical-url-functionality
Adds the ability to pass a new parameter canonicalUrl in the front matter of a post that is used to generate a canonical link in the head of the page. This will enable us to post blogs that originally appeared elsewhere with the proper attribution so that search engines do not penalize the site for duplicate content.

Signed-off-by: Erik Bledsoe <erik@calyptia.com>