Add canonical link functionality (#34)
* add-canonical-url-functionality

Adds the ability to pass a new parameter canonicalUrl in the front matter of a post that is used to generate a canonical link in the head of the page. This will enable us to post blogs that originally appeared elsewhere with the proper attribution so that search engines do not penalize the site for duplicate content.

Signed-off-by: Erik Bledsoe <erik@calyptia.com>

* Create fluent-bit-surpasses-10-billion-docker-pulls.md

Adds most recent Fluent Bit blog post from calyptia.com

Signed-off-by: Erik Bledsoe <erik@calyptia.com>

* Update fluent-bit-surpasses-10-billion-docker-pulls.md

Fix formatting. Add canonicalUrl to front matter.

Signed-off-by: Erik Bledsoe <erik@calyptia.com>

* fix url

Signed-off-by: Erik Bledsoe <erik@calyptia.com>

---------

Signed-off-by: Erik Bledsoe <erik@calyptia.com>