Update _index.md
This commit updates the description of Fluent Bit to reflect its ability to process traces in addition to logging and metrics. 

The original line mentioned only logging and metrics, so the updated line now highlights Fluent Bit's capability to handle logs, metrics, and traces, aligning the description with Fluent Bit's full feature set.

Signed-off-by: Sampath Siva Kumar Boddeti <119833411+sampathshivakumar@users.noreply.github.com>